pulled functions out into a module
